Normalize WHOOP BLE sensor data to standard physical units
IMU data was stored as raw int16 LSB values — meaningless without knowing
the sensor configuration and incompatible with Apple Watch/iPhone data
which is already in standard units (g for accel, rad/s for gyro).

- Accelerometer: raw int16 / 4096 → g (confirmed ±8g range, 4096 LSB/g)
- Gyroscope: raw int16 / 16.4 × π/180 → rad/s (assumed ±2000 dps)
- WhoopImuSample fields changed from Int16 to Float
- JS bridge serializes as Double instead of Int
- Removed rawPayload from WhoopRealtimeDataSample (not stored in DB)
- Added test for gravity normalization (raw 4096 → 1.0g)

+++ packages/mobile/modules/whoop-ble/ios/WhoopBleModule.swift
@@ -377,12 +377,12 @@ public class WhoopBleModule: Module {
 
                 return [
                     "timestamp": formatter.string(from: date),
-                    "accelerometerX": Int(sample.accelerometerX),
-                    "accelerometerY": Int(sample.accelerometerY),
-                    "accelerometerZ": Int(sample.accelerometerZ),
-                    "gyroscopeX": Int(sample.gyroscopeX),
-                    "gyroscopeY": Int(sample.gyroscopeY),
-                    "gyroscopeZ": Int(sample.gyroscopeZ),
+                    "accelerometerX": Double(sample.accelerometerX),
+                    "accelerometerY": Double(sample.accelerometerY),
+                    "accelerometerZ": Double(sample.accelerometerZ),
+                    "gyroscopeX": Double(sample.gyroscopeX),
+                    "gyroscopeY": Double(sample.gyroscopeY),
+                    "gyroscopeZ": Double(sample.gyroscopeZ),
                 ]
             }
